save the simulated data points as png
