+++ simulator/src/config.yaml
@@ -1,43 +1,46 @@
+---
+cluster_name: test
 cluster_hostname: test-cluster.aws
 cluster_ip_address: 10.0.0.1
-cluster_name: test
-data_generation_interval_minutes: 5
-data_ingestion:
-  randomness_percentage: 10
-  states:
-  - ingestion_rate_gb_per_hr: 10
-    position: 1
-    time_hh_mm_ss: '00_00_00'
-  - ingestion_rate_gb_per_hr: 60
-    position: 2
-    time_hh_mm_ss: '06_00_00'
-  - ingestion_rate_gb_per_hr: 65
-    position: 3
-    time_hh_mm_ss: '12_00_00'
-  - ingestion_rate_gb_per_hr: 15
-    position: 4
-    time_hh_mm_ss: '18_00_00'
-  - ingestion_rate_gb_per_hr: 5
-    position: 5
-    time_hh_mm_ss: '23_00_00'
-data_nodes_count: 4
-index_clean_up_age_days: 20
-index_count: 1000
-index_roll_over_size: 50
-master_eligible_nodes_count: 1
 node_machine_type_identifier: m5-12xlarge
-searches:
-- cpu_load_percent: 10
-  memory_load_percent: 10
-  probability: 0.3
-  type: simple
-- cpu_load_percent: 20
-  memory_load_percent: 20
-  probability: 0.25
-  type: medium
-- cpu_load_percent: 30
-  memory_load_percent: 30
-  probability: 0.15
-  type: complex
-shards_per_index: 5
 total_nodes_count: 5
+data_nodes_count: 4
+master_eligible_nodes_count: 1
+index_count: 1000
+shards_per_index: 5
+index_roll_over_size: 50
+index_clean_up_age_days: 20
+data_generation_interval_minutes: 5
+
+data_ingestion:
+  states:
+  - position: 1
+    time_hh_mm_ss: '00_00_00'
+    ingestion_rate_gb_per_hr: 10
+  - position: 2
+    time_hh_mm_ss: '06_00_00'
+    ingestion_rate_gb_per_hr: 60
+  - position: 3
+    time_hh_mm_ss: '12_00_00'
+    ingestion_rate_gb_per_hr: 65
+  - position: 4
+    time_hh_mm_ss: '18_00_00'
+    ingestion_rate_gb_per_hr: 15
+  - position: 5
+    time_hh_mm_ss: '23_00_00'
+    ingestion_rate_gb_per_hr: 5
+  randomness_percentage: 5
+
+searches:
+- type: simple
+  probability: 0.3
+  cpu_load_percent: 10
+  memory_load_percent: 10
+- type: medium
+  probability: 0.25
+  cpu_load_percent: 20
+  memory_load_percent: 20
+- type: complex
+  probability: 0.15
+  cpu_load_percent: 30
+  memory_load_percent: 30
