+++ simulator/src/constants.py
@@ -1,5 +1,17 @@
 
-SCHEMA_FILE_NAME = 'schema.py' # Todo : Check if needed
+# file paths
+SCHEMA_FILE_NAME = 'schema.py'
+CONFIG_FILE_PATH = 'config.yaml'
+PROVISION_LOCK_FILE_NAME = 'provision.lock'
+SIMULATION_GRAPH_FILE_NAME = 'simulation_diagram.png'
+
+# config keys
+DATA_INGESTION = "data_ingestion"
+DATA_INGESTION_STATES = 'states'
+SEARCHES = "searches"
+DATA_GENERATION_INTERVAL_MINUTES = "data_generation_interval_minutes"
+DATA_INGESTION_RANDOMNESS_PERCENTAGE = 'randomness_percentage'
+# Todo - Shrinidhi/Manoj add remaining keys from config yaml
 
 CLUSTER_STATE_GREEN = 'green'
 CLUSTER_STATE_YELLOW = 'yellow'
@@ -9,6 +21,7 @@ CLUSTER_STATES = [CLUSTER_STATE_GREEN, CLUSTER_STATE_YELLOW, CLUSTER_STATE_RED]
 
 HIGH_INGESTION_RATE_GB_PER_HOUR = 60
 
+# mapping inputs for API endpoints
 CLUSTER_STATE = 'status'
 CPU_USAGE_PERCENT = 'cpu_usage_percent'
 MEMORY_USAGE_PERCENT = 'memory_usage_percent'
@@ -17,14 +30,6 @@ STAT_REQUEST = {
     'mem': MEMORY_USAGE_PERCENT,
     'status': CLUSTER_STATE
 
-} # Todo : Add remaining stats  
+}  # Todo : Shrinidhi/Manoj Add remaining stats
 
 APP_PORT = 5000
-
-CONFIG_PATH = 'config.yaml'
-
-DATA_INGESTION = "data_ingestion"
-SEARCHES = "searches"
-DATA_GENERATION_INTERVAL_MINUTES = "data_generation_interval_minutes"
-
-PROVISION_LOCK_FILE_NAME = 'provision.lock'
